add faults tests for the multi stage query
