+++ extensions-core/multi-stage-query/src/test/java/org/apache/druid/msq/exec/MSQInsertTest.java
@@ -27,12 +27,10 @@ import com.google.common.hash.Hashing;
 import org.apache.druid.hll.HyperLogLogCollector;
 import org.apache.druid.java.util.common.ISE;
 import org.apache.druid.java.util.common.Intervals;
-import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.java.util.common.granularity.Granularities;
 import org.apache.druid.msq.indexing.error.ColumnNameRestrictedFault;
 import org.apache.druid.msq.indexing.error.InsertTimeNullFault;
 import org.apache.druid.msq.indexing.error.RowTooLargeFault;
-import org.apache.druid.msq.indexing.error.TooManyClusteredByColumnsFault;
 import org.apache.druid.msq.test.MSQTestBase;
 import org.apache.druid.msq.util.MultiStageQueryContext;
 import org.apache.druid.query.aggregation.LongSumAggregatorFactory;
@@ -55,8 +53,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 public class MSQInsertTest extends MSQTestBase
 {
@@ -448,48 +444,6 @@ public class MSQInsertTest extends MSQTestBase
                      .verifyPlanningErrors();
   }
 
-  @Test
-  public void testInsertWithHugeClusteringKeys()
-  {
-    RowSignature dummyRowSignature = RowSignature.builder().add("__time", ColumnType.LONG).build();
-
-    final int numColumns = 1700;
-
-    String columnNames = IntStream.range(1, numColumns)
-                                  .mapToObj(i -> "col" + i).collect(Collectors.joining(", "));
-
-    String clusteredByClause = IntStream.range(1, numColumns + 1)
-                                        .mapToObj(String::valueOf)
-                                        .collect(Collectors.joining(", "));
-
-    String externSignature = IntStream.range(1, numColumns)
-                                      .mapToObj(i -> StringUtils.format(
-                                          "{\"name\": \"col%d\", \"type\": \"string\"}",
-                                          i
-                                      ))
-                                      .collect(Collectors.joining(", "));
-
-    testIngestQuery()
-        .setSql(StringUtils.format(
-            " insert into foo1 SELECT\n"
-            + "  floor(TIME_PARSE(\"timestamp\") to day) AS __time,\n"
-            + " %s\n"
-            + "FROM TABLE(\n"
-            + "  EXTERN(\n"
-            + "    '{ \"files\": [\"ignored\"],\"type\":\"local\"}',\n"
-            + "    '{\"type\": \"json\"}',\n"
-            + "    '[{\"name\": \"timestamp\", \"type\": \"string\"}, %s]'\n"
-            + "  )\n"
-            + ") PARTITIONED by day CLUSTERED BY %s",
-            columnNames,
-            externSignature,
-            clusteredByClause
-        ))
-        .setExpectedDataSource("foo1")
-        .setExpectedRowSignature(dummyRowSignature)
-        .setExpectedMSQFault(new TooManyClusteredByColumnsFault(numColumns + 2, 1500, 0))
-        .verifyResults();
-  }
 
   @Test
   public void testInsertRestrictedColumns()
